#hot-fix: complete rename queryConstants -> messageQueryConstants
